Logging refactor
