+++ example/out_example.cpp
@@ -1,4 +1,5 @@
 #include <matching.hpp>
+#include <logging.hpp>
 #include <fstream>
 #include <sstream>
 #include <string>
@@ -6,34 +7,13 @@
 using namespace std;
 
 
-void writeHierarchyToFile(const urquhart::Observation& trees, std::string fileName) {
+void writeHierarchyFiles(const urquhart::Observation& trees, std::string fileName) {
     std::ofstream plyOut(fileName+"-p.txt"), triOut(fileName+"-t.txt"), 
                     hieOut(fileName+"-h.txt"), dscOut(fileName+"-d.txt");
-    trees.hier->viewTree(hieOut);
-    hieOut.close();
     
-    // Iterate over the indices of the Polygons in the hierarchy
-    for(auto pIdx : trees.hier->getChildrenIds(0)) {
-        for (int i = 0; i < trees.hier->getPolygon(pIdx).landmarkRefs.size(); ++i) {
-            auto myPoint =  trees.landmarks.col(trees.hier->getPolygon(pIdx).landmarkRefs(i));
-            plyOut << pIdx << " " << myPoint[0] << " " << myPoint[1] << "|";
-        }
-        plyOut << std::endl;
-        for(auto d : trees.hier->getPolygon(pIdx).descriptor) dscOut << d << " ";
-        dscOut << std::endl;
-        
-        // Iterate over the indices of the Triangles that compose this Polygon
-        for(auto tIdx : trees.hier->getChildrenIds(pIdx)) {
-            // Retain only the Polygon objects that have three sides
-            if (trees.hier->getPolygon(tIdx).n == 3) {
-                for (int i = 0; i < trees.hier->getPolygon(tIdx).landmarkRefs.size(); ++i) {
-                    auto myPoint = trees.landmarks.col(trees.hier->getPolygon(tIdx).landmarkRefs(i));
-                    triOut << tIdx << " " << myPoint[0] << " " << myPoint[1] << "|";
-                }
-                triOut << std::endl;
-            }
-        }
-    }
+    logging::writeHierarchyToFile(trees, plyOut, triOut, hieOut, dscOut);
+    
+    hieOut.close();
     plyOut.close();
     triOut.close();
     dscOut.close();
@@ -77,7 +57,7 @@ int main(int argc, char* argv[]) {
         urquhart::Observation obs1(points1);
         std::cout << "Obs 1 Polygons: " << std::endl;
         obs1.view();
-        writeHierarchyToFile(obs1, "1"+num);
+        writeHierarchyFiles(obs1, "1"+num);
 
         std::ofstream pts1Out("1"+num+"-pts.txt");
         for (int i = 0; i < obs1.landmarks.cols(); ++i) {
@@ -89,7 +69,7 @@ int main(int argc, char* argv[]) {
         urquhart::Observation obs2(points2);
         std::cout << "Obs 2 Polygons: " << std::endl;
         obs2.view();
-        writeHierarchyToFile(obs2, "2"+num);
+        writeHierarchyFiles(obs2, "2"+num);
 
         std::ofstream pts2Out("2"+num+"-pts.txt");
         for (int i = 0; i < obs2.landmarks.cols(); ++i) {
