+++ include/urquhart/distance.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <fstream>
 #include <iostream>
 #include <vector>
 #include <algorithm>
